options for wp-docker commands, repository url, keep .git directory (defaults to false). Code cleanup.

+++ packages/cli/cmd/docker.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
-const util = require('util');
-const exec = util.promisify(require('child_process').exec);
+const util = require("util");
+const exec = util.promisify(require("child_process").exec);
 const prompts = require("prompts");
 const fse = require("fs-extra");
 const path = require("path");
@@ -8,6 +8,8 @@ const figlet = require("figlet");
 const chalk = require("chalk");
 const replace = require("replace-in-file");
 const { features } = require("process");
+const { program } = require("commander");
+const { options } = require("args");
 
 const PROCESS_PATH = path.join(process.cwd());
 
@@ -17,17 +19,26 @@ let cancelled = false;
 const REPOSITORY_URL = "git@github.com:Mill3/wordpress-docker-boilerplate.git";
 const DEFAULT_DOCKER_PORT_WEB = 8000;
 const DEFAULT_DOCKER_PORT_PHPMYADMIN = 8001;
-const DEFAULT_DOCKER_MASTER_PATH = '/Users/myuser/wp-install-root/';
+const DEFAULT_DOCKER_MASTER_PATH = "/Users/myuser/wp-install-root/";
+const DEFAULT_THEME_NAME = "mytheme";
 
 // defaults
 let settings = {
+  REPOSITORY_URL: REPOSITORY_URL,
   DOCKER_PORT_WEB: DEFAULT_DOCKER_PORT_WEB,
   DOCKER_PORT_PHPMYADMIN: DEFAULT_DOCKER_PORT_PHPMYADMIN,
   INSTALL_PATH: PROCESS_PATH,
+  THEME_NAME: DEFAULT_THEME_NAME
 };
 
 // start Prompts
 const run = async () => {
+  const options = program.opts();
+
+  if (options.repository) {
+    settings["REPOSITORY_URL"] = options.repository;
+  }
+
   // First Intro message
   console.log(chalk.blue("*****************************************************"));
   console.log(chalk.blue(figlet.textSync("MILL3", { horizontalLayout: "full" })));
@@ -62,6 +73,15 @@ const run = async () => {
       }
     },
     {
+      type: "text",
+      name: "THEME_NAME",
+      initial: settings["THEME_NAME"],
+      message: "What is your theme directory name (my-theme, my-project-theme-directory, my_super_theme) ?",
+      validate: text => {
+        return /^[A-Za-z0-9]+(?:[-_][A-Za-z0-9]+)*$/.test(text) ? true : "In slug-format only.";
+      }
+    },
+    {
       type: "toggle",
       name: "confirm",
       message: `Are you sure? This will overwrite any same exiting file in target directory !`,
@@ -93,21 +113,32 @@ const run = async () => {
   console.log(chalk.blue(`\nDone with questions, starting installation in directory ${settings["INSTALL_PATH"]}\n`));
 
   // install all files
-  await install();
+  await install(options);
 
   // rename to new namespace
-  await rename();
+  await rename(options);
 
   console.log(chalk.blue(`\nInstallation done in ${settings["INSTALL_PATH"]}\n`));
 };
 
 // clone and cleanup files
-const install = async () => {
+const install = async options => {
   try {
     console.log(chalk.green(`Step 1 : cloning repository in directory ${settings["INSTALL_PATH"]}`));
-    await clone();
-    await fse.remove(`${settings["INSTALL_PATH"]}/.git`);
-    await fse.move(`${settings["INSTALL_PATH"]}/.env.sample`, `${settings["INSTALL_PATH"]}/.env`, { overwrite: true });
+
+    // clone repos
+    const { error, stdout, stderr } = await exec(`git clone ${settings["REPOSITORY_URL"]} ${settings["INSTALL_PATH"]}`);
+    if (error) {
+      throw error;
+    } else {
+      console.log(chalk.green(`Step 2 : repo cloning done.`));
+    }
+
+    // remove git directory if specified
+    if (!options.keepgit) await fse.remove(`${settings["INSTALL_PATH"]}/.git`);
+
+    // rename .env
+    await fse.copy(`${settings["INSTALL_PATH"]}/.env.sample`, `${settings["INSTALL_PATH"]}/.env`, { overwrite: false });
   } catch (err) {
     // display Git error
     console.error(err);
@@ -116,27 +147,24 @@ const install = async () => {
   }
 };
 
-const clone = async () => {
-  const { error, stdout, stderr } = await exec(`git clone ${REPOSITORY_URL} ${settings["INSTALL_PATH"]}`);
-  if (error) {
-    throw error;
-  } else {
-    console.log(chalk.green(`Step 2 : repo cloning done.`));
-  }
-}
-
 const rename = async () => {
   // regex for domain and namespace
   var reDockerWeb = new RegExp(DEFAULT_DOCKER_PORT_WEB, "g");
   var reDockerPhpMyAdmin = new RegExp(DEFAULT_DOCKER_PORT_PHPMYADMIN, "g");
   var reDockerMasterPath = new RegExp(DEFAULT_DOCKER_MASTER_PATH, "g");
+  var reDockerThemeName = new RegExp(DEFAULT_THEME_NAME, "g");
 
   // replace options
   const options = {
     files: [`${settings["INSTALL_PATH"]}/.env`],
     ignore: ["node_modules/**"],
-    from: [reDockerWeb, reDockerPhpMyAdmin, reDockerMasterPath],
-    to: [settings["DOCKER_PORT_WEB"], settings["DOCKER_PORT_PHPMYADMIN"], settings["INSTALL_PATH"]],
+    from: [reDockerWeb, reDockerPhpMyAdmin, reDockerMasterPath, reDockerThemeName],
+    to: [
+      settings["DOCKER_PORT_WEB"],
+      settings["DOCKER_PORT_PHPMYADMIN"],
+      settings["INSTALL_PATH"],
+      settings["THEME_NAME"]
+    ],
     countMatches: true
   };
 
@@ -147,7 +175,6 @@ const rename = async () => {
     console.error(err);
   }
 };
-
 
 run();
